Improve drawing

+++ libmatrix/drawing.h
@@ -9,7 +9,7 @@ typedef struct pixel_s {
 } pixel;
 
 void picture_line(picture_t * pic, pixel start, pixel end, int on);
-void picture_circle(picture_t * pic, pixel center, unsigned int radius,
+void picture_circle(picture_t * pic, pixel center, double radius,
 		    unsigned int on);
 
 #endif
